removed dependency of sourcery on sparsedetect.
