+++ suite2p/detection/sourcery.py
@@ -5,12 +5,12 @@ import numpy as np
 from scipy.ndimage import filters
 from scipy.ndimage import gaussian_filter
 
-from . import sparsedetect
+from .utils import bin_movie, get_sdmov
 from .. import utils
 
 
 def getSVDdata(ops):
-    mov, max_proj = sparsedetect.bin_movie(ops)
+    mov, max_proj = bin_movie(ops)
     ops['max_proj'] = max_proj
     nbins, Lyc, Lxc = np.shape(mov)
 
@@ -19,7 +19,7 @@ def getSVDdata(ops):
         mov[j,:,:] = gaussian_filter(mov[j,:,:], sig)
 
     # compute noise variance across frames
-    sdmov = sparsedetect.get_sdmov(mov, ops)
+    sdmov = get_sdmov(mov, ops)
     mov /= sdmov
     mov = np.reshape(mov, (-1,Lyc*Lxc))
     if 1:
@@ -40,7 +40,7 @@ def getSVDdata(ops):
     return ops, U, sdmov, u
 
 def getSVDproj(ops, u):
-    mov, _ = sparsedetect.bin_movie(ops)
+    mov, _ = bin_movie(ops)
 
     nbins, Lyc, Lxc = np.shape(mov)
     if ('smooth_masks' in ops) and ops['smooth_masks']:
@@ -49,7 +49,7 @@ def getSVDproj(ops, u):
             mov[j,:,:] = gaussian_filter(mov[j,:,:], sig)
     if 1:
         #sdmov = np.ones((Lyc*Lxc,), 'float32')
-        sdmov = sparsedetect.get_sdmov(mov, ops)
+        sdmov = get_sdmov(mov, ops)
         mov/=sdmov
         mov = np.reshape(mov, (-1,Lyc*Lxc))
